Refactor gpt-tfjs to support config as intit argument
